break up module.rb
