+++ lib/app.rb
@@ -21,7 +21,7 @@ require_relative 'app/utils/logging'
 require_relative 'app/endpoint'
 require_relative 'app/utils/secure_random_base62'
 require_relative 'app/sequence_base'
-require_relative 'app/module'
+require_relative 'app/models/module'
 require_relative 'version'
 require_relative 'app/models'
 require_relative 'app/utils/terminology'
