+++ test/test_helper.rb
@@ -14,7 +14,6 @@ require 'json/jwt'
 test_log_filename = File.join('tmp', 'test.log')
 FileUtils.rm test_log_filename if File.exist? test_log_filename
 
-# require File.expand_path '../../app.rb', __FILE__
 require_relative '../lib/app'
 
 def load_json_fixture(file)
